Move `with`-style testing helpers to a separate file
